Create MathOperations.py
This is one of my Python's scripts I've ever write. It's a math problem with makes random operations between random numbers . If the answer it's wrong, the loop continues and asking you the wright answer again and again and count the wrong answers. If the answer it's correct goes further to the next operation based on number of operations you've declared as a global variable. In the end the script will return the time that takes you to answer to the math problems and the number of wrong answers.